requirements.txt
Flask==3.0.3
openpyxl==3.1.5
Werkzeug==3.0.3
gunicorn==23.0.0
python-dotenv==1.0.0
